A lot of bug fixes.
Previous version was returning a spreadsheet which was too big (square in size)
due to the list comprehension approach, replaced with zip, and incorrect usage of split
which was leaving the separators in the spreadsheet.

Added argument to denote if the function should take the names of the coulmns from the csv
or generate default names.

Defined what functions are exposed in the module.

Added a filter to remove empty columns contained in the csv.

Added temporary import testing tools.
Will be removed onece import is working flawless.

TODO: There is one glitch connected to having the , sign as value in a column.
      In this case it is writen as "," in the csv, but importCSV will read it
      as two " signs separetad with , . Maybe playing around with oneOf can fix the problem.

	  Also - what if the file is completely empty.